Registering and logging in
"But are you really there? (Are you really there?)
Are you really there? (Are you really there?)"

Added a controller that infterfaces with database in order to create users and check if users exist

Added a user DTO to seperate database and the API which users will interface with

Added a service which will generate JWTs

Modified Program.cs to support authentication and authorization.

Added a key to appsettings.json. Probably should hide it though....