Fixing up webscraper + adding user outfit post and get
UserDataController now has functionality to create an outfit associated with a user, along with retrieving it, both using JWT information.

Webscraper needed updates as source changed their website slightly!

Nuked migrations, as functionality for user was combined with the clothinitemcontext.

Updated naming of outfits.cs to outfit, since it is a singular item after all.